+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [0.1.0]
+* Subscriptions Create
 * Users Examples 1 - 5 https://docs.microsoft.com/en-us/graph/api/user-list
 * Initial Version